SortList refactoring
