+++ src/components/PageDefinitions.js
@@ -3,24 +3,15 @@
 
 import React from 'react'
 import { connect } from 'react-redux'
-import { Grid, Button, DropdownButton, MenuItem, OverlayTrigger, Tooltip } from 'react-bootstrap'
 import pako from 'pako'
-
+import { Grid, Button, DropdownButton, MenuItem, OverlayTrigger, Tooltip } from 'react-bootstrap'
 import base64js from 'base64-js'
-import { saveAs } from 'file-saver'
 import notification from 'antd/lib/notification'
-
-import trim from 'lodash/trim'
 import get from 'lodash/get'
-import { uiNavigation, uiBrowseUpdateList, uiRevertDefinition, uiInfo, uiWarning, uiDanger } from '../actions/ui'
-import { getDefinitionsAction } from '../actions/definitionActions'
-import { ROUTE_CURATIONS, ROUTE_DEFINITIONS, ROUTE_SHARE } from '../utils/routingConstants'
-import EntitySpec from '../utils/entitySpec'
+import { uiNavigation, uiBrowseUpdateList, uiWarning } from '../actions/ui'
+import { ROUTE_DEFINITIONS } from '../utils/routingConstants'
 import AbstractPageDefinitions from './AbstractPageDefinitions'
-import { getCurationAction } from '../actions/curationActions'
 import NotificationButtons from './Navigation/Ui/NotificationButtons'
-import { asObject } from '../utils/utils'
-import { getGist, saveGist } from '../api/github'
 import SearchBar from './Navigation/Ui/SearchBar'
 import { ComponentList, Section, ContributePrompt } from './'
 import FullDetailPage from './FullDetailView/FullDetailPage'
@@ -166,245 +157,6 @@ export class PageDefinitions extends AbstractPageDefinitions {
           </ul>
         </div>
       </div>
-    )
-  }
-
-  doSave() {
-    const { components } = this.props
-    const spec = this.buildSaveSpec(components.list)
-    this.saveSpec(spec)
-    this.setState({ showSavePopup: false, fileName: null })
-  }
-
-  // capture this work in a serpate method so it can be spun off without waiting but still capture rejections
-  async saveSpec(spec) {
-    const { dispatch } = this.props
-    try {
-      const fileObject = { filter: this.state.activeFilters, sortBy: this.state.activeSort, coordinates: spec }
-      if (this.state.saveType === 'gist') await this.createGist(this.state.fileName, fileObject)
-      else {
-        const file = new File([JSON.stringify(fileObject, null, 2)], `${this.state.fileName}.json`)
-        saveAs(file)
-      }
-    } catch (error) {
-      if (error.status === 404)
-        return uiWarning(dispatch, "Could not create Gist. Likely you've not given us permission")
-      uiWarning(dispatch, error.message)
-    }
-  }
-
-  async createGist(name, content) {
-    const { token, dispatch } = this.props
-    const url = await saveGist(token, `${name}.json`, JSON.stringify(content))
-    const message = (
-      <div>
-        A new Gist File has been created and is available{' '}
-        <a href={url} target="_blank" rel="noopener noreferrer">
-          here
-        </a>
-      </div>
-    )
-    return uiInfo(dispatch, message)
-  }
-
-  doSaveAsUrl() {
-    const { components } = this.props
-    const spec = this.buildSaveSpec(components.list)
-    const fileObject = { filter: this.state.activeFilters, sortBy: this.state.activeSort, coordinates: spec }
-    const url = `${document.location.origin}${ROUTE_SHARE}/${base64js.fromByteArray(
-      pako.deflate(JSON.stringify(fileObject))
-    )}`
-    this.copyToClipboard(url, 'URL copied to clipboard')
-  }
-
-  copyToClipboard(text, message) {
-    const textArea = document.createElement('textarea')
-    textArea.value = text
-    document.body.appendChild(textArea)
-    textArea.focus()
-    textArea.select()
-    document.execCommand('copy')
-    document.body.removeChild(textArea)
-    uiInfo(this.props.dispatch, message)
-  }
-
-  onDragOver = e => e.preventDefault()
-  onDragEnter = e => e.preventDefault()
-
-  onDrop = async e => {
-    e.preventDefault()
-    e.persist()
-    try {
-      if ((await this.handleTextDrop(e)) !== false) return
-      if (this.handleDropFiles(e) !== false) return
-      uiWarning(this.props.dispatch, 'ClearlyDefined does not understand whatever it is you just dropped')
-    } catch (error) {
-      uiWarning(this.props.dispatch, error.message)
-    }
-  }
-
-  handleTextDrop = async event => {
-    const text = event.dataTransfer.getData('Text')
-    if (!text) return false
-    if (this.handleDropObject(text) !== false) return
-    if ((await this.handleDropGist(text)) !== false) return
-    if (this.handleDropEntityUrl(text) !== false) return
-    if (this.handleDropPrURL(text) !== false) return
-    return false
-  }
-
-  // handle dropping a URL to an npm, github repo/release, nuget package, ...
-  handleDropEntityUrl(content) {
-    const spec = EntitySpec.fromUrl(content)
-    if (!spec) return false
-    this.onAddComponent(spec)
-  }
-
-  // dropping an actual definition, an object that has `coordinates`
-  handleDropObject(content) {
-    const contentObject = asObject(content)
-    if (!contentObject) return false
-    this.onAddComponent(EntitySpec.fromCoordinates(contentObject))
-  }
-
-  // handle dropping a url pointing to a curation PR
-  handleDropPrURL(urlSpec) {
-    try {
-      const url = new URL(trim(urlSpec, '/'))
-      if (url.hostname !== 'github.com') return false
-      const [, org, , type, number] = url.pathname.split('/')
-      if (org !== 'clearlydefined' || type !== 'pull') return false
-      this.props.history.push(`${ROUTE_CURATIONS}/${number}`)
-    } catch (exception) {
-      return false
-    }
-  }
-
-  // handle dropping a url to a Gist that contains a ClearlyDefined coordinate list
-  async handleDropGist(urlString) {
-    if (!urlString.startsWith('https://gist.github.com')) return false
-    uiInfo(this.props.dispatch, 'Loading component list from gist')
-    const url = new URL(urlString)
-    const [, , id] = url.pathname.split('/')
-    if (!id) throw new Error(`Gist url ${url} is malformed`)
-    const content = await getGist(id)
-    if (!content || !Object.keys(content).length) throw new Error(`Gist at ${url} could not be loaded or was empty`)
-    for (let name in content) this.loadComponentList(content[name], name)
-  }
-
-  handleDropFiles(event) {
-    const files = Object.values(event.dataTransfer.files)
-    if (!files || !files.length) return false
-    const { acceptedFiles, rejectedFiles } = this.sortDroppedFiles(files)
-    if (acceptedFiles.length) this.handleDropAcceptedFiles(acceptedFiles)
-    if (rejectedFiles.length) this.handleDropRejectedFiles(rejectedFiles)
-  }
-
-  sortDroppedFiles(files) {
-    const acceptedFilesValues = ['application/json']
-    return files.reduce(
-      (result, file) => {
-        if (acceptedFilesValues.includes(file.type)) result.acceptedFiles.push(file)
-        else result.rejectedFiles.push(file)
-        return result
-      },
-      { acceptedFiles: [], rejectedFiles: [] }
-    )
-  }
-
-  handleDropAcceptedFiles(files) {
-    uiInfo(this.props.dispatch, 'Loading component list from file(s)')
-    files.forEach(file => {
-      const reader = new FileReader()
-      reader.onload = () => this.loadComponentList(reader.result, file.name)
-      reader.readAsBinaryString(file)
-    })
-  }
-
-  handleDropRejectedFiles = files => {
-    const fileNames = files.map(file => file.name).join(', ')
-    uiWarning(this.props.dispatch, `Could not load: ${fileNames}`)
-  }
-
-  onAddComponent(value) {
-    const { dispatch, token, definitions } = this.props
-    const component = typeof value === 'string' ? EntitySpec.fromPath(value) : value
-    const path = component.toPath()
-    if (!component.revision) return uiWarning(dispatch, `${path} needs version information`)
-
-    !definitions.entries[path] &&
-      dispatch(getDefinitionsAction(token, [path])) &&
-      dispatch(getCurationAction(token, component))
-    dispatch(uiBrowseUpdateList({ add: component }))
-  }
-
-  dropZone(child) {
-    return (
-      <div
-        onDragOver={this.onDragOver}
-        onDragEnter={this.onDragEnter}
-        onDrop={this.onDrop}
-        style={{ position: 'relative' }}
-      >
-        {child}
-      </div>
-    )
-  }
-
-  loadComponentList(content, name) {
-    const list = this.getList(content)
-    if (!list) return uiWarning(this.props.dispatch, `Invalid component list file: ${name}`)
-    this.loadFromListSpec(list)
-  }
-
-  getList(content) {
-    const object = typeof content === 'string' ? JSON.parse(content) : content
-    if (this.isPackageLock(object)) return this.getListFromPackageLock(object.dependencies)
-    if (this.isClearlyDefinedList(object)) return object
-    return null
-  }
-
-  isPackageLock(content) {
-    // TODO better, more definitive test here
-    return !!content.dependencies
-  }
-
-  isClearlyDefinedList(content) {
-    // TODO better, more definitive test here
-    return !!content.coordinates
-  }
-
-  getListFromPackageLock(dependencies) {
-    const coordinates = []
-    for (const dependency in dependencies) {
-      let [namespace, name] = dependency.split('/')
-      if (!name) {
-        name = namespace
-        namespace = null
-      }
-      coordinates.push({ type: 'npm', provider: 'npmjs', namespace, name, revision: dependencies[dependency].version })
-    }
-    return { coordinates }
-  }
-
-  loadFromListSpec(list) {
-    const { dispatch, definitions } = this.props
-    if (list.filter) this.setState({ activeFilters: list.filter })
-    if (list.sortBy) this.setState({ activeSort: list.sortBy })
-    if (list.sortBy || list.filter) this.setState({ sequence: this.state.sequence + 1 })
-
-    const toAdd = list.coordinates.map(component => EntitySpec.validateAndCreate(component)).filter(e => e)
-    dispatch(uiBrowseUpdateList({ addAll: toAdd }))
-    const missingDefinitions = toAdd.map(spec => spec.toPath()).filter(path => !definitions.entries[path])
-    this.getDefinitionsAndNotify(missingDefinitions, 'All components have been loaded')
-    dispatch(
-      uiBrowseUpdateList({
-        transform: this.createTransform.call(
-          this,
-          list.sortBy || this.state.activeSort,
-          list.filter || this.state.activeFilters
-        )
-      })
     )
   }
